ia-dashboard-backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Data processing
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4

# AI providers
google-generativeai==0.7.2
openai==1.35.7

# Database
supabase==2.5.0
sqlalchemy==2.0.30
asyncpg==0.29.0

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# PDF export
reportlab==4.2.2

# Config and validation
pydantic[email]==2.7.3
pydantic-settings==2.3.1
python-dotenv==1.0.1

# Logging
structlog==24.2.0

# HTTP client (for AI provider calls)
httpx==0.27.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
factory-boy==3.3.0

# Dev tools
ruff==0.4.9

# Others
psycopg2-binary